+++ package.json
@@ -52,5 +52,8 @@
     "eslint": "^5.16.0",
     "eslint-plugin-import": "^2.17.2",
     "nyc": "^14.1.1"
+  },
+  "dependencies": {
+    "json-parser": "^2.1.1"
   }
 }
