+++ src/stringify.js
@@ -2,10 +2,10 @@
 
 module.exports = stringify
 
-/////////////////////////////////////////////////////////////////////////////////////////////////////////
+/////////////////////////////////////////////////////////////////
 // Modified from Douglas Crockford's JSON2:
 // https://github.com/douglascrockford/JSON-js
-/////////////////////////////////////////////////////////////////////////////////////////////////////////
+/////////////////////////////////////////////////////////////////
 
 const OBJ_PROTO = Object.prototype
 const KEY_PREFIX = '// '
